Modularizar arquivos e testes, substituindo Karma por jasmine node
