+++ .travis.yml
@@ -1,9 +1,5 @@
 language: node_js
 node_js:
     - "lts/*"
-script: node_modules/karma/bin/karma start my.conf.js
-before_install:
-    - export DISPLAY=:99.0
-    - sh -e /etc/init.d/xvfb start
 before_script:
-    - npm install
+    - npm i -g
